+++ keyboards/wilba_tech/wt_main.c
@@ -16,6 +16,152 @@
 
 #include "quantum.h"
 #include "keyboards/wilba_tech/wt_mono_backlight.h"
+#include "keyboards/zeal60/zeal60_api.h" // Temporary hack
+
+#include "raw_hid.h"
+#include "dynamic_keymap.h"
+#include "timer.h"
+#include "tmk_core/common/eeprom.h"
+
+bool eeprom_is_valid(void)
+{
+	return (eeprom_read_word(((void*)EEPROM_MAGIC_ADDR)) == EEPROM_MAGIC &&
+			eeprom_read_byte(((void*)EEPROM_VERSION_ADDR)) == EEPROM_VERSION);
+}
+
+void eeprom_set_valid(bool valid)
+{
+	eeprom_update_word(((void*)EEPROM_MAGIC_ADDR), valid ? EEPROM_MAGIC : 0xFFFF);
+	eeprom_update_byte(((void*)EEPROM_VERSION_ADDR), valid ? EEPROM_VERSION : 0xFF);
+}
+
+void eeprom_reset(void)
+{
+	// Set the Zeal60 specific EEPROM state as invalid.
+	eeprom_set_valid(false);
+	// Set the TMK/QMK EEPROM state as invalid.
+	eeconfig_disable();
+}
+
+#ifdef RAW_ENABLE
+
+void raw_hid_receive( uint8_t *data, uint8_t length )
+{
+	uint8_t *command_id = &(data[0]);
+	uint8_t *command_data = &(data[1]);
+	switch ( *command_id )
+	{
+		case id_get_protocol_version:
+		{
+			command_data[0] = PROTOCOL_VERSION >> 8;
+			command_data[1] = PROTOCOL_VERSION & 0xFF;
+			break;
+		}
+		case id_get_keyboard_value:
+		{
+			if ( command_data[0] == id_uptime )
+			{
+				uint32_t value = timer_read32();
+				command_data[1] = (value >> 24 ) & 0xFF;
+				command_data[2] = (value >> 16 ) & 0xFF;
+				command_data[3] = (value >> 8 ) & 0xFF;
+				command_data[4] = value & 0xFF;
+			}
+			else
+			{
+				*command_id = id_unhandled;
+			}
+			break;
+		}
+#ifdef DYNAMIC_KEYMAP_ENABLE
+		case id_dynamic_keymap_get_keycode:
+		{
+			uint16_t keycode = dynamic_keymap_get_keycode( command_data[0], command_data[1], command_data[2] );
+			command_data[3] = keycode >> 8;
+			command_data[4] = keycode & 0xFF;
+			break;
+		}
+		case id_dynamic_keymap_set_keycode:
+		{
+			dynamic_keymap_set_keycode( command_data[0], command_data[1], command_data[2], ( command_data[3] << 8 ) | command_data[4] );
+			break;
+		}
+		case id_dynamic_keymap_reset:
+		{
+			dynamic_keymap_reset();
+			break;
+		}
+#endif // DYNAMIC_KEYMAP_ENABLE
+		case id_backlight_config_set_value:
+		{
+			//backlight_config_set_value(command_data);
+			break;
+		}
+		case id_backlight_config_get_value:
+		{
+			//backlight_config_get_value(command_data);
+			break;
+		}
+		case id_backlight_config_save:
+		{
+			//backlight_config_save();
+			break;
+		}
+		case id_eeprom_reset:
+		{
+			eeprom_reset();
+			break;
+		}
+		case id_bootloader_jump:
+		{
+			// Need to send data back before the jump
+			// Informs host that the command is handled
+			raw_hid_send( data, length );
+			// Give host time to read it
+			wait_ms(100);
+			bootloader_jump();
+			break;
+		}
+		default:
+		{
+			// Unhandled message.
+			*command_id = id_unhandled;
+			break;
+		}
+	}
+
+	// Return same buffer with values changed
+	raw_hid_send( data, length );
+
+}
+
+#endif
+
+void main_init(void)
+{
+	// If the EEPROM has the magic, the data is good.
+	// OK to load from EEPROM.
+	if (eeprom_is_valid()) {
+		//backlight_config_load();
+	} else	{
+		// If the EEPROM has not been saved before, or is out of date,
+		// save the default values to the EEPROM. Default values
+		// come from construction of the zeal_backlight_config instance.
+		//backlight_config_save();
+#ifdef DYNAMIC_KEYMAP_ENABLE
+		// This resets the keymaps in EEPROM to what is in flash.
+		dynamic_keymap_reset();
+#endif
+		// Save the magic number last, in case saving was interrupted
+		eeprom_set_valid(true);
+	}
+
+	// Initialize LED drivers for backlight.
+	backlight_init_drivers();
+
+	backlight_timer_init();
+	backlight_timer_enable();
+}
 
 void bootmagic_lite(void)
 {
@@ -32,9 +178,7 @@ void bootmagic_lite(void)
 	// If the Esc (matrix 0,0) is held down on power up,
 	// reset the EEPROM valid state and jump to bootloader.
 	if ( matrix_get_row(0) & (1<<0) ) {
-		// Set the TMK/QMK EEPROM state as invalid.
-		eeconfig_disable();
-		// Jump to bootloader.
+		eeprom_reset();
 		bootloader_jump();
 	}
 }
@@ -42,14 +186,13 @@ void bootmagic_lite(void)
 void matrix_init_kb(void)
 {
 	bootmagic_lite();
-	backlight_init_drivers();
-	backlight_timer_init();
-	backlight_timer_enable();
+	main_init();
 	matrix_init_user();
 }
 
 void matrix_scan_kb(void)
 {
+	// This only updates the LED driver buffers if something has changed.
 	backlight_update_pwm_buffers();
 	matrix_scan_user();
 }
