+++ keyboards/zeal65/config.h
@@ -79,10 +79,10 @@
 
 // enable/disable LEDs based on layout
 #define RGB_BACKLIGHT_USE_SPLIT_BACKSPACE 0
-#define RGB_BACKLIGHT_USE_SPLIT_LEFT_SHIFT 0
-#define RGB_BACKLIGHT_USE_SPLIT_RIGHT_SHIFT 1
-#define RGB_BACKLIGHT_USE_7U_SPACEBAR 0
-#define RGB_BACKLIGHT_USE_ISO_ENTER 0
+#define RGB_BACKLIGHT_USE_SPLIT_LEFT_SHIFT 0  // Unused
+#define RGB_BACKLIGHT_USE_SPLIT_RIGHT_SHIFT 0 // Unused
+#define RGB_BACKLIGHT_USE_7U_SPACEBAR 1       // Used only for stab LED matching spacebar LED
+#define RGB_BACKLIGHT_USE_ISO_ENTER 0         // Unused
 #define RGB_BACKLIGHT_DISABLE_HHKB_BLOCKER_LEDS 0
 
 // disable backlight when USB suspended (PC sleep/hibernate/shutdown)
